fixes getting tasks and versions for steps
